A simple loader for the platformer creator I made with Godot
And yeah, the idea is extremely inspired by THAT title.

The loader works and has been tested, however this only load
the level main blocks.
The architecture of the save is still undefined at that point.
I wanted to be sure that the idea could be implemented, and
what would be the issues.

The next steps will be :
* Avoid using a Mesh Collider on loading, and generate
  multiple 1x1 Cube colliders instead, see if that has
  any positive impact.
* Change the save architecture to have one block ID and
  all the positions of that block, in order to avoid
  searching for the same block multiple times on generation.
* Change the save architecture to provide information
  about collectibles, interactive elements, enemies and
  goals.
  (The basics of a platformer)

Signed-off-by: Voyage <voyage@miouyouyou.fr>